# broker: subscribers: use a slice instead of a map