--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,3 @@
 module github.com/peterbourgon/ps
 
 go 1.20
-
-require github.com/bernerdschaefer/eventsource v0.0.0-20130606115634-220e99a79763