In summary, using a slice makes Publish faster by up to 1 order of magnitude, but also makes Subscribe-Unsubscribe slower by order(s) of magnitude proportional to the number of active subscribers. I think this is a net win, because Publish performance matters a lot more than Subscribe-Unsubscribe performance. Thanks to @oliverpool for the tip in #1.

```
name                                      old time/op    new time/op    delta
BrokerMethods/Publish/0-11                  7.17ns ± 0%    6.17ns ± 3%    -13.90%  (p=0.008 n=5+5)
BrokerMethods/Publish/1-11                  34.3ns ± 0%     7.6ns ± 1%    -77.86%  (p=0.016 n=4+5)
BrokerMethods/Publish/100-11                1.05µs ± 4%    0.33µs ± 0%    -68.26%  (p=0.008 n=5+5)
BrokerMethods/Publish/10000-11               108µs ± 1%      33µs ± 0%    -69.87%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/0-11                78.6ns ± 1%    81.3ns ± 0%     +3.39%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/1-11                78.1ns ± 1%    82.6ns ± 1%     +5.76%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/100-11              95.8ns ± 4%   195.1ns ± 1%   +103.70%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/10000-11             109ns ± 2%    9958ns ± 2%  +9027.77%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/0-11        35.1ns ± 2%     6.5ns ± 1%    -81.55%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/1-11        96.3ns ± 4%    16.5ns ± 0%    -82.91%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/100-11      1.45µs ± 2%    0.27µs ± 2%    -81.47%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/10000-11     183µs ± 0%      28µs ± 2%    -84.91%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
BrokerMethods/Publish/0-11                   0.00B          0.00B            ~     (all equal)
BrokerMethods/Publish/1-11                   0.00B          0.00B            ~     (all equal)
BrokerMethods/Publish/100-11                 0.00B          0.00B            ~     (all equal)
BrokerMethods/Publish/10000-11               0.00B          0.00B            ~     (all equal)
BrokerMethods/Sub-Unsub/0-11                  144B ± 0%      160B ± 0%    +11.11%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/1-11                  144B ± 0%      160B ± 0%    +11.11%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/100-11                144B ± 0%      160B ± 0%    +11.11%  (p=0.008 n=5+5)
BrokerMethods/Sub-Unsub/10000-11              144B ± 0%      160B ± 0%    +11.11%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/0-11         24.0B ± 0%      0.0B        -100.00%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/1-11         80.0B ± 0%     24.0B ± 0%    -70.00%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/100-11      6.25kB ± 0%    2.69kB ± 0%    -56.98%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/10000-11     574kB ± 0%     246kB ± 0%    -57.15%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
BrokerMethods/Publish/0-11                    0.00           0.00            ~     (all equal)
BrokerMethods/Publish/1-11                    0.00           0.00            ~     (all equal)
BrokerMethods/Publish/100-11                  0.00           0.00            ~     (all equal)
BrokerMethods/Publish/10000-11                0.00           0.00            ~     (all equal)
BrokerMethods/Sub-Unsub/0-11                  3.00 ± 0%      3.00 ± 0%       ~     (all equal)
BrokerMethods/Sub-Unsub/1-11                  3.00 ± 0%      3.00 ± 0%       ~     (all equal)
BrokerMethods/Sub-Unsub/100-11                3.00 ± 0%      3.00 ± 0%       ~     (all equal)
BrokerMethods/Sub-Unsub/10000-11              3.00 ± 0%      3.00 ± 0%       ~     (all equal)
BrokerMethods/ActiveSubscribers/0-11          1.00 ± 0%      0.00        -100.00%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/1-11          3.00 ± 0%      1.00 ± 0%    -66.67%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/100-11        5.00 ± 0%      1.00 ± 0%    -80.00%  (p=0.008 n=5+5)
BrokerMethods/ActiveSubscribers/10000-11      5.00 ± 0%      1.00 ± 0%    -80.00%  (p=0.008 n=5+5)
```